factor Memfd out from CircularBuffer
The reason for doing this is so we don't have to store the fd handle in the CircularBuffer struct.  This removes a detail about CircularBuffer that probably shouldn't be abstracted because the caller may want to know that a memfd has been created and they may want to control if/when/how it is closed.